streamlit-app/requirements.txt
streamlit
pandas
google-cloud-bigquery
